feat(taplo-lsp): import #:schema-directives relative to file
The `#:schema` comment directive is a useful way of clearly describing
the schema that is used to check a toml file. Since taplo hasn’t
publicly documented it yet, we still have the ability to change its
functionality in backwards-incompatible ways.

This commit changes the logic, which resolved relative
schema-directives relative to the workspace root, which taplo
determines once at the startup of the extension.

This has the problem that multi-folder workspaces do not resolve it
correctly, as documented in
https://github.com/tamasfe/taplo/issues/118

It also has the problem: if you follow the relative link given in the
file, you will only find the schema if the file’s directory is the
workspace directory. So “go to file” does only work if the file is in
the workspace toplevel, not if the file is anywhere else.

Instead, we resolve *relative to the toml file*. This is the
“intuitive” way, and not any harder to implement, in fact it’s less
stateful and trivially works.

As noted in the comment, we could also implement it for e.g. `http`
urls, since the `Url` type supports relative `.join()` on any
resource. But that will require some more work.