fix: remove all the array index keys when there is more than one
The `Query::position_info_at()` function can return one index when the
value is a non-empty array. In this case, the `Query::lookup_keys()`
function will add another one, resulting in two index keys. But the code
was only removing one, causing hover to break for keys with a non-empty
array value.

This change avoids the issue by removing index keys in a while loop.

Closes #788